Fix all the nixpkgs update fallout (#85)
* Vendor `cuda-packages` from nixpkgs

nixpkgs has removed CUDA 11.8 and 12.0 to 12.5. However, in
kernel-builder we need:

- CUDA 11.8 for Torch 2.7.
- CUDA 12.4 for FA3 (currently errors out on newer nvcc).

So, vendor old CUDA versions, at least for the time being.

* Vendor nixpkgs gcc derivation for gcc 11 and 12

gcc11Stdenv is required for CUDA 11.8. We are using gcc12Stdenv in
kernel-builder for ROCm kernel builds.

* python3Packages.triton-rocm: 3.2.0 -> 3.4.0

Also add patches to fix the build.

* python3Packages.torch_2_7: fix Triton 2.4 compatibility

* python3Packages.torch_2_6: remove

* Scrub maintainers

We do not want to bother upstream maintainers with our vendored
packages.